Added getting usersFavoritedItineraryList to the HomePage now just need to manipulate the itinerary data to display appropriately.

+++ src/main/java/com/example/AdventureAppraisals/Controllers/UserFavoritesController.java
@@ -73,7 +73,7 @@ public class UserFavoritesController {
                 List<Integer> newUserFavoritedItineraryList = userFavorite.getFavoritedItineraryList();
 
                 if (newUserFavoritedItineraryList.contains(itinerary.getId())) {
-                    newUserFavoritedItineraryList.remove(itinerary.getId());
+                    newUserFavoritedItineraryList.remove((Integer) itinerary.getId());
                 } else {
                     return null;
                 }
